베터리 통신 및 pose 핑키 통신 추가 빌드 잘됨

+++ Service/ROS_Task_Manager/robocallee_fms/Core/include/Commondefine.hpp
@@ -167,7 +167,7 @@ namespace Commondefine
     typedef struct RobotTaskInfo
     {
         int                                 robot_id;
-        int                                 battery = 100;
+        float                               battery = 100.f;
         shoesproperty                       shoes_property;
         pose2f                              dest;
         std::string                         requester;
